pyproject.toml
[project]
name = "mainkode"
version = "0.1.0"
description = ""
authors = [
    {name = "shrestic",email = "nhatphongcgp@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.11"

dependencies = [
    "agate (==1.9.1)",
    "annotated-types (==0.7.0)",
    "appdirs (==1.4.4)",
    "asn1crypto (==1.5.1)",
    "atpublic (==5.0)",
    "attrs (==24.2.0)",
    "babel (==2.16.0)",
    "backports-tarfile (==1.2.0)",
    "boto3 (==1.35.45)",
    "botocore (==1.35.45)",
    "cachetools (==5.5.0)",
    "certifi (==2024.8.30)",
    "cffi (==1.17.1)",
    "chardet (==5.2.0)",
    "charset-normalizer (==3.4.0)",
    "click (==8.1.7)",
    "colorama (==0.4.6)",
    "cryptography (==43.0.1)",
    "daff (==1.3.46)",
    "dbt-adapters (==1.7.0)",
    "dbt-common (==1.11.0)",
    "dbt-core (==1.8.7)",
    "dbt-extractor (==0.5.1)",
    "dbt-semantic-interfaces (==0.5.1)",
    "dbt-snowflake (==1.8.3)",
    "deepdiff (==7.0.1)",
    "diff-cover (==9.2.0)",
    "exceptiongroup (==1.2.2)",
    "filelock (==3.16.1)",
    "fire (==0.7.0)",
    "httplib2 (==0.22.0)",
    "idna (==3.10)",
    "importlib-metadata (==6.11.0)",
    "iniconfig (==2.0.0)",
    "isodate (==0.6.1)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jinja2 (==3.1.4)",
    "jinja2-simple-tags (==0.6.1)",
    "jmespath (==1.0.1)",
    "jsonschema (==4.23.0)",
    "jsonschema-specifications (==2024.10.1)",
    "keyring (==25.4.1)",
    "leather (==0.4.0)",
    "logbook (==1.5.3)",
    "markupsafe (==3.0.1)",
    "mashumaro (==3.13.1)",
    "minimal-snowplow-tracker (==0.0.2)",
    "more-itertools (==10.5.0)",
    "msgpack (==1.1.0)",
    "networkx (==3.2.1)",
    "numpy (==2.0.2)",
    "oauthlib (==3.2.2)",
    "ordered-set (==4.1.0)",
    "packaging (==24.1)",
    "pandas (==2.2.3)",
    "parsedatetime (==2.6)",
    "pathspec (==0.12.1)",
    "platformdirs (==4.3.6)",
    "pluggy (==1.5.0)",
    "proto-plus (==1.24.0)",
    "protobuf (==4.25.5)",
    "psycopg2-binary (==2.9.10)",
    "pyarrow (==17.0.0)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.1)",
    "pycparser (==2.22)",
    "pydantic (==2.9.2)",
    "pydantic-core (==2.23.4)",
    "pygments (==2.18.0)",
    "pygsheets (==2.0.5)",
    "pyjwt (==2.9.0)",
    "pyopenssl (==24.2.1)",
    "pyparsing (==3.2.0)",
    "pytest (==8.3.3)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "python-slugify (==8.0.4)",
    "pytimeparse (==1.1.8)",
    "pytz (==2024.2)",
    "pyyaml (==6.0)",
    "referencing (==0.35.1)",
    "regex (==2024.9.11)",
    "requests (==2.32.3)",
    "requests-oauthlib (==2.0.0)",
    "rpds-py (==0.20.0)",
    "rsa (==4.9)",
    "s3transfer (==0.10.3)",
    "six (==1.16.0)",
    "snowflake (==0.13.1)",
    "snowflake-connector-python (==3.12.3)",
    "snowflake-sqlalchemy (==1.6.1)",
    "snowflake-core (==0.13.1)",
    "sortedcontainers (==2.4.0)",
    "sqlalchemy (==1.4.36)",
    "sqlfluff (==3.2.5)",
    "sqlfluff-templater-dbt (==3.2.5)",
    "sqlparse (==0.5.1)",
    "tblib (==3.0.0)",
    "termcolor (==2.5.0)",
    "text-unidecode (==1.3)",
    "toml (==0.10.2)",
    "tomli (==2.0.2)",
    "tomlkit (==0.13.2)",
    "tqdm (==4.66.5)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2024.2)",
    "uritemplate (==4.1.1)",
    "urllib3 (==1.26.20)",
    "zipp (==3.20.2)",
    "apache-airflow (==2.10.0)",
    "apache-airflow-providers-cncf-kubernetes (==10.1.0)",
    "pytest-mock (==3.12.0)",
    "pytest-responses (==0.5.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false
[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"
[tool.black]
line-length = 88  # Ensure it matches your desired line length (usually black's default)
[tool.isort]
profile = "black"  # Ensure isort follows black's import formatting style
line_length = 88  # Ensure line length matches black